+++ R/data.r
@@ -1,7 +1,8 @@
 #' Flea beatle measurements
 #'
 #' This data is from a paper by A. A. Lubischew, "On the Use of Discriminant
-#' Functions in Taxonomy", Biometrics, Dec 1962, pp.455-477.
+#' Functions in Taxonomy", Biometrics, Dec 1962, pp.455-477. Data is
+#' standardized, and original units are in flea_raw.
 #'
 #' \itemize{
 #'   \item tars1, width of the first joint of the first tarsus in microns
@@ -16,7 +17,7 @@
 #' }
 #'
 #' @name Flea measurements
-#' @aliases flea
+#' @aliases flea, flea_raw
 #' @docType data
 #' @format A 74 x 7 numeric array
 #' @keywords datasets
@@ -25,7 +26,7 @@
 #' head(flea)
 #' animate_xy(flea[, -7])
 #' animate_xy(flea[, -7], col = flea[, 7])
-NULL
+"flea"
 
 #' Turnable laser measurements from Bellcore
 #'