+++ R/interesting-indices.r
@@ -103,8 +103,8 @@ loess2d <- function() {
 #'
 #' Compares the similarity between the projected distribution and a normal distribution.
 #' \itemize{
-#' \item{norm_bin }{compares the count in 100 histogram bins}
-#' \item{norm_kol }{compares the cdf based on the Kolmogorov–Smirnov test (KS test)}
+#' \item norm_bin: compares the count in 100 histogram bins
+#' \item norm_kol: compares the cdf based on the Kolmogorov–Smirnov test (KS test)
 #' }
 #' @param nr The number of rows in the target matrix
 #'